Kafka/kafka-python-hello/requirements.txt
bpython
flake8
kafka-python
mypy
pronounceable
pylint
